Add option to regenerate missing world rather than discarding config.
Adds `regenerate-missing-worlds` to `settings.yml`, with the default being `false` (keeping behavior in-line with previous versions).

This should allow for niche use cases, i.e a test server environment with config mirrored from production. Generating a new world in-place with the same configuration allows for server developers to make stand-in test worlds for their production counterparts.

Without this option, tracking and potentially losing a world's configuration due to not manually regenerating a new test world becomes a nightmare.